Move dusk modules to separate files
